this doesn't work, ill fix tmmrw

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/vision/simulator/pipeline1Simulator.java
@@ -131,7 +131,8 @@ public class pipeline1Simulator extends OpenCvPipeline {
         return input;
     }
 
-    private int mostDifferent(double box1, double box2, double box3) {
+    private byte mostDifferent(double box1, double box2, double box3) {
+        //this is broken, i'll fix tmrmw
         double boxMean = (box1+box2+box3)/3;
         double array[] = {Math.abs(boxMean - box1),Math.abs(boxMean - box2),Math.abs(boxMean - box3)};
         double max = 0;
